remove fp16 std support from float16.h
