+++ cinn/common/type.h
@@ -19,6 +19,7 @@
 #include <string>
 
 #include "cinn/common/float16.h"
+#include "cinn/common/float16_utils.h"
 #include "cinn/common/macros.h"
 #include "cinn/runtime/cinn_runtime.h"
