+++ cinn/runtime/custom_function_test.cc
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
-
 #include <gtest/gtest.h>
 
 #include <functional>
